Create new module queue_executor
